Add support for `@javax.annotation.CheckReturnValue`
Since immutable objects are stateless and the invocation of methods on
immutable objects does not have observable side effects, one should generally
use the result of such invocations. A class-level `@CheckReturnValue`
annotation communicates this fact. Tools such as Error Prone can check for
adherence to this contract [1].

An exception to this rule are various methods defined on the nested Builder
class, as those methods return `this`. Those methods may thus be annotated
`@com.google.errorprone.annotations.CanIgnoreReturnValue`.

This change only applies `@CheckReturnValue` if _both_ `@CheckReturnValue` and
`@CanIgnoreReturnValue` are on the classpath, for application of the former
without the latter needlessly forces users of the nested builder into a fluent
usage style.

[1] http://errorprone.info/bugpattern/CheckReturnValue