Call copyOf on add methods if deepImmutablesDetection is enabled
Fix for #333.

In situations like this:

    @Value.Style(deepImmutablesDetection = true)
    @Value.Immutable
    @Value.Modifiable
    public interface Foo {
        List<Bar> getBars();

        // ...
    }

    @Value.Immutable
    @Value.Modifiable
    public interface Bar {
        // ...
    }

When the following methods are called:
 ImmutableFoo.Builder#bars(Iterable<Bar>),
 ImmutableFoo.Builder#addBar(Bar),
 ImmutableFoo.Builder#addBar(Bar ...),
 ImmutableFoo.Builder#addAllBar(Iterable<Bar>)
 ImmutableFoo.Builder#from(Foo)

then the Bar parameters will be converted with ImmutableBar.copyOf
before being added to the list. This is more consistent with
non-collection @Value fields when deepImmutablesDetection is on, and
prevents cases where you have an ImmutableFoo with ModifiableBars
inside.