Call Modifiable factory on elements in add methods if deepImmutablesDetection is enabled
In situations like this:

    @Value.Style(deepImmutablesDetection = true)
    @Value.Immutable
    @Value.Modifiable
    public interface Foo {
        List<Bar> getBars();

        // ...
    }

    @Value.Immutable
    @Value.Modifiable
    public interface Bar {
        // ...
    }

When the following methods are called:
 ModifiableFoo#setBars(Iterable<Bar>),
 ModifiableFoo#addBar(Bar),
 ModifiableFoo#addBar(Bar ...),
 ModifiableFoo#addAllBar(Iterable<Bar>)
 ModifiableFoo#from(Foo)

then the Bar parameters will be converted with

ModifiablePoint.create().from()

before being added to the list if they are not already ModifiablePoints.
This is more consistent with non-collection @Value fields when
deepImmutablesDetection is on, and prevents cases where you have an
ModifiableFoo with ImmutableBars inside.