Create proper standalone jar for metainf processor
The metainf project produces a standalone jar, but the pom file still
declares the shaded dependencies which defeats the purpose of creating a
shaded standalone jar.

By aligning the metainf pom file with the value pom file a proper
standalone jar is created for the metainf processor. This removes the
standalone classifier and creates a standalone jar by default.