Fix nondeterminism in code generation of nested values (#677)
Because `HashSet` iteration order can vary from run to run, the order in which the processor
was generating nested value classes could change from run to run. This causes problems with
build systems like Buck that assume deterministic output so that artifacts can be cached.

Switching to a `LinkedHashSet` takes care of the issue.