Support protocol buffer attributes
The discovery mechanism needs to be more strict for protocol buffers
because there are several methods which have the right signature, but
are not part of the builder API.

First, we make sure all methods for an attribute builder match the
specified Value#attributeBuilder patterns

Second, we ensure that the return methods are equivalent using
typeMirriors#isSameType. Using equals was working intermittently, and
the documentation on equals says to use the isSameType instead.