Explicitly fail for repositories with Gson polymorphic mapping (@ExpectedSubtypes)
Since migration to [mongo codecs](http://mongodb.github.io/mongo-java-driver/3.10/bson/codecs/) a case
of polymorphic serialization was missed (gson serialization). This seems to be an edge case and it is safer to fail fast
during repository code generation until a proper solution is in place.

For more information see #934

Closes #934