Make generated MongoDB repositories extensible (#948)
* Make generated MongoDB repositories extensible

We have use cases in which we want to aggregate documents of generated repositories. Since the generator does not provide an API to build an Aggregation Pipeline, we write them by hand. In this case, we want to extend the generated repository with custom functionality where we need access to the used MongoCollection.

* Refinement