Add legacy SourceOrdering option
Prior to Immutables 2.7.5, the source ordering was
defined be a "shallow" approach where accessors from
the child class were added first, then the superclass,
then the interfaces. This changed in 2.7.5, where the
order became interfaces, then superclass, then child
class. Issue 889 acknowledged this as a breaking
change, and it has made it difficult for some clients
to migrate safely to later versions.

Add a new system property,
org.immutables.useLegacyAccessorOrdering,
that when set to true, reverts to the <=2.7.4 ordering.
This should enable more clients to migrate to more
recent versions safely.