Optimize mergeInternal generation with specialized methods per subtype
Prior to this commit, our mergeInternal function that is created when
a derived class is used as an Immutables target would check the possible
types in a series of if blocks. However, this can lead to an explosion
in the length of this method, especially for target classes with multiple
parents. Long methods are concerning because it can exceed the JVM's
default 8 KB limit for method compilation, forcing calls to remain
interpreted.

Instead, generate specialized methods for each relevant subtype. This
helps break up the monolithic mergeInternal method while not introducing
too many method hops.

This approach, however, requires special handling of our bit masks,
since we may have more than one long value to represent cases where the
model has >64 fields. To deal with this, introduce a specialized case
where we use a long[] of values, in order for each specialized method to
easily update the overall state. This approach balances optimizing the
simple case (<= 64 fields) for performance on small models, while
allowing for larger models with only introducing a minimal amount of
overhead in the allocation and access of the array values.